Rearranged objects directory.
Moved ship to its own directory for simpler organization.

Moved base objects to objects/__init__.py, which required a change of the imports for all objects.

Fixed a client side bug when no port exists.
