--- game/objects/base.py
@@ -1,56 +1,0 @@
-import yaml
-import uuid
-from copy import deepcopy
-
-class GameObject(yaml.YAMLObject):
-    __metaclass__ = yaml.YAMLObjectMetaclass
-
-    def __init__(self):
-        super(GameObject,self).__init__()
-        # Call byteify to make sure all unicode variables are saved as strings
-        # This makes it easier to save in yaml
-        self.__dict__ = self.byteify(self.__dict__)
-
-    def byteify(self, data):
-        """Make sure everything in data is a string, not unicode (for nicer yaml)."""
-        if isinstance(data, dict):
-            result = {}
-            for key, value in data.iteritems():
-                if isinstance(key, unicode):
-                    key = key.encode('utf-8')
-                if isinstance(value, unicode):
-                    value = value.encode('utf-8')
-                if isinstance(value, dict) or isinstance(value, list):
-                    value = self.byteify(value)
-                result[key] = value
-        elif isinstance(data, list):
-            result = []
-            for item in data:
-                if isinstance(item, unicode):
-                    item = item.encode('utf-8')
-                if isinstance(item, dict) or isinstance(item, list):
-                    item = self.byteify(value)
-                result.append(item)
-        else:
-            result = None
-
-        return result
-
-    def __repr__(self):
-        return str(self.__dict__)
-
-    # This method is used by json.dumps() when sending it to a client
-    def __str__(self):
-        return str(self.__dict__)
-
-    def plural(self, capitalized = False):
-        """Return the plural form of this object."""
-        return self.__class__.__name__.lower() + "s"
-
-    def to_dict(self):
-        return deepcopy(self.__dict__)
-
-    @classmethod
-    def from_yaml(cls, loader, node):
-        fields = loader.construct_mapping(node, deep=True)
-        return cls(**fields)
