+++ game/objects/planet.py
@@ -1,7 +1,7 @@
 import uuid
-from coordinates import Coordinates
 from random import choice
-from sector import SectorObject
+from objects.sector import SectorObject
+from objects.coordinates import Coordinates
 
 class Planet(SectorObject):
     yaml_tag = "!Planet"
